// src/shared/api/dcim/generated/models/DiskConfigurationDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Medialand API
 * Medialand API description
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
/**
 *
 * @export
 * @interface DiskConfigurationDto
 */
export interface DiskConfigurationDto {
  /**
   *
   * @type {string}
   * @memberof DiskConfigurationDto
   */
  type: DiskConfigurationDtoTypeEnum;
  /**
   *
   * @type {number}
   * @memberof DiskConfigurationDto
   */
  volume: number;
}

export function DiskConfigurationDtoFromJSON(json: any): DiskConfigurationDto {
  return DiskConfigurationDtoFromJSONTyped(json, false);
}

export function DiskConfigurationDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): DiskConfigurationDto {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    type: json["type"],
    volume: json["volume"]
  };
}

export function DiskConfigurationDtoToJSON(value?: DiskConfigurationDto | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    type: value.type === undefined ? null : value.type,
    volume: value.volume === undefined ? null : value.volume
  };
}

/**
 * @export
 * @enum {string}
 */
export enum DiskConfigurationDtoTypeEnum {
  SSD = "SSD",
  NvME = "NvME",
  HDD = "HDD"
}